Discord event notifications.
